requirements.txt
annotated-types==0.7.0
docopt==0.6.2
grpcio==1.71.0
grpcio-tools==1.71.0
protobuf==5.29.4
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
python-dotenv==1.1.0
setuptools==79.0.1
stomp.py==8.2.0
typing-inspection==0.4.0
typing_extensions==4.13.2
websocket-client==1.8.0
